Reliable UART interrupt-driven Tx/Rx test passes
UART transmit and receive interrupt-driven code now working reliably. Tx code primes the transmitter and hands control to the ISR. When a transmit message is exhausted it is restarted in the background main loop. Rx interrupts are continuously battered in the foreground by continuous transmission from RealTerm running on a seperate laptop.