Implementing free buffer handling
While the UART code is still being developed it useful to loop the receiver to the transmitter. Rather than just send out received characters one-by-one separate transmit and receive ring-buffers mimic the intended independent transmit/receive operation. For a loop when a full Rx ring-buffer arrives it can be immediately assigned to the transmitter - in a queue itself comprised of a ring-buffer. The receiver then needs a new supply of ring-buffers so a "free" list is used where ring-buffers are pulled first to the receiver, then to the transmitter and finally back to the "free" list.