Interrupt-driven loop-back
Here the receive and transmit primary serial UART are fully interrupt-driven. A circulating "free" list of ring-buffers is maintained on a list, to be pulled by the receiver, transferred to the transmitter when full and returned to the "free" list when exhausted. At 19200bps full-chat the transmitter cannot keep up with the receiver and is eventually overwhelmed. This is not of any great concern as the purpose of this build is just to do some stress-tests on the primary serial UART code which is holding up very well. There is no intention to use this as a high-speed loopback, only an intermittent command processor. In any case if push came to shove the peripheral DMA is a better way to go. Experiment is a success.