Decoupled tx and rx at the primary UART
So far it has been convenient to stress the code by running masses of characters in the receiver and looping then around to the transmitter. The loop-back is still there but now it is in the background code. When a receiver ring-buffer fills it is placed onto an independent ring of ring-buffers. This is in preparation for the primary command protocol task.