Test and Control
The primary command and control (during development) to the Arduino is a serial UART. The Microsoft serial service via their ".ocx" is anecdotally not very good (I have used it myself in the past and it does seem a bit flaky). The "RealTerm" (see "https://realterm.sourceforge.io/") comms tool works very well as a manual interface. As it happens the RealTerm team have also implemented automation via COM/ActiveX. Using COM/ActiveX from C is not well documented on the web - lots of simple C++ examples and hidden libraries for the lesser scripting languages. Luckily there is an excellent library at "http://disphelper.sourceforge.net/". Some work is required between the RealTerm examples (the Matlab example is very useful - if you have Matlab) and the "disphelper" library to figure out how to drive it all together. The files here are the first example of firing up RealTerm, doing some configuration and continuously driving characters down the serial port. All written in VS 12.0 'C' (incredibly it also compiles and runs on Vista under VS 2005!). The version of RealTerm to use is 3.0.0.27.