Messaging layer code started
The message handling in APV is split between low-level code such as the serial UART interrupt service routine, low-level message assembly and verification and a message handling layer. In principle the message handling layer is not limited to hardware-in/firmware-out, the handling functions can be "stacked" as ALL messaging between communications handlers uses a common message buffer structure. In addition non-communications processes can access the communications I/O by injecting into the messaging layer via the common message buffers. Message transfer is atomic at the message level so ad-hoc streaming of single characters is not supported but ad-hoc streaming of whole messages is. All messages are embedded in a defined message structure for identification and verification purposes so picking relevant messages out of the stream should not be a problem. This mechanism is intended for control and low-speed data NOT continuous streaming.