EXPERIMENTAL Duration Timers
As an experiment moved the duration timer function triggered on SysTick from the background loop into the SysTick ISR to guarantee the timing interval. Some caution is required here since this is a "fast" interrupt currently running at 150usecs - not a lot of room for large callback functions! Best used just to set flags or drive strobes.